Find optimal solutions for all status-code conversion macros
Guarantied optimal because rather than (continuing to) figure these out by hand, I've wrote a brute-force script to do a breadth-first search to find the shortest sequence of operations that satisfies the mapping requirements.

+++ src/deemon/method-hints/seq_operator_getitem.h
@@ -756,7 +756,7 @@ err:
 	                                 (void *)(uintptr_t)index, index, end_index);
 	ASSERT(foreach_status == -4 || foreach_status == -3 || foreach_status == -2 ||
 	       foreach_status == -1 || foreach_status == 0);
-	return Dee_BOUND_FROM_eM1_m0_mM4_yM2_nM3(foreach_status);
+	return Dee_BOUND_FROM_eM1_m0_mM4_yM2_nM3((int)foreach_status);
 }} = $with__seq_operator_getitem_index;
 
 seq_operator_bounditem_index = {
