+++ src/deemon/runtime/method-hint-defaults.c
@@ -1948,7 +1948,7 @@ default__seq_operator_bounditem_index__with__seq_enumerate_index(DeeObject *__re
 	foreach_status = (*DeeType_RequireMethodHint(Dee_TYPE(self), seq_enumerate_index))(self, &default_bounditem_index_with_seq_enumerate_index_cb, (void *)(uintptr_t)index, index, end_index);
 	ASSERT(foreach_status == -4 || foreach_status == -3 || foreach_status == -2 ||
 	       foreach_status == -1 || foreach_status == 0);
-	return Dee_BOUND_FROM_eM1_m0_mM4_yM2_nM3(foreach_status);
+	return Dee_BOUND_FROM_eM1_m0_mM4_yM2_nM3((int)foreach_status);
 }
